fixing some small issues from #15
